requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-groq
langgraph
sqlalchemy
alembic
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
faiss-cpu
chromadb
openai
groq
requests
aiohttp
numpy
pandas
psycopg2-binary